requirements.txt
fastapi[standard]
pydantic>=2
requests
fastmcp            # Modern MCP library for FastAPI integration
python-dotenv      # For loading environment variables from .env
